chore: test refactor
